Update msgFBmodal.js
Updated the document so that the table is not used anymore.
Also replaced all the words fbmodel with the word boxmodel. Did this because facebook is a commercial bussiness and doesn't like if you use that kind of name.

The border is now visible with transparency and can be set in the CSS file.